pyproject.toml
[project]
name = "cci-design"
version = "0.1.0"
description = "CCI Timesheet Reporting System - Python automation for generating weekly/monthly timesheet reports and invoices"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.1",
    "caldav>=2.1.2",
    "faker>=39.0.0",
    "fastapi[standard]>=0.128.0",
    "msal>=1.34.0",
    "msgraph-sdk>=1.51.0",
    "numbers-parser>=4.16.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
